LPS-56986 Simplify/SF
