+++ modules/sdk/ant-bnd/test/unit/com/liferay/ant/bnd/spring/SampleBean.java
@@ -12,22 +12,14 @@
  * details.
  */
 
-package com.liferay.ant.bnd.spring.annotation;
-
-import java.lang.annotation.Documented;
-import java.lang.annotation.ElementType;
-import java.lang.annotation.Retention;
-import java.lang.annotation.RetentionPolicy;
-import java.lang.annotation.Target;
+package com.liferay.ant.bnd.spring;
 
 /**
  * @author Miguel Pastor
  */
-@Documented
-@Retention(RetentionPolicy.RUNTIME)
-@Target({ElementType.FIELD, ElementType.METHOD})
-public @interface ServiceReference {
+public class SampleBean {
 
-	public Class<?> type() default Object.class;
+	@ServiceReference(type = String.class)
+	protected String s;
 
 }
